Refactor app

+++ crawler/app/main.py
@@ -1,6 +1,11 @@
 import datetime
+import warnings
+from time import sleep
 import shutil
 import subprocess
+import os
+
+import psycopg2
 from PyPDF2 import PdfFileReader
 
 from io import StringIO
@@ -29,13 +34,35 @@ def extract_pdf(root: Path):
 
 
 if __name__ == '__main__':
-    connection = pg.connect(dbname='postgres',
-                            user='postgres',
-                            host='localhost',
-                            port=5432,
-                            password='sdcovid')
+    try:
+        db_host = os.environ['DB_HOST']
+    except KeyError:
+        db_host = 'localhost'
+
+    retries = 0
+    print('Trying to connect to database')
+    while True:
+        try:
+            connection = pg.connect(dbname='postgres',
+                                    user='postgres',
+                                    host=db_host,
+                                    port=5432,
+                                    password='sdcovid')
+        except psycopg2.OperationalError as e:
+            sleep(10)
+            if retries==10:
+                raise ConnectionError('Could not connect to database.')
+            else:
+                warnings.warn('Error when connecting to database: {}'.format(str(e)))
+            retries += 1
+        else:
+            print('Connection successful.')
+            break
+
     with connection as conn:
+        print('Check if table exists...')
         if not check_if_table_exists(conn, SQL_TABLE_NAME):
+            print('Creating table since it does not exist.')
             with conn.cursor() as cur:
                 cur.execute("""
                     CREATE TABLE cases_by_zip 
@@ -47,50 +74,65 @@ if __name__ == '__main__':
                         );
                 """)
                 conn.commit()
+        else:
+            print('Table exists.')
+    print('Done checking table.')
 
     data_dir = Path(__file__).parent.joinpath('./data')
+    skip_download = False
     if data_dir.exists():
-        shutil.rmtree(data_dir)
-    data_dir.mkdir()
+        if 'INIT_DB' in os.environ and any(data_dir.iterdir()):
+            skip_download = True
+        else:
+            shutil.rmtree(data_dir)
+    data_dir.mkdir(exist_ok=True)
 
     with connection as conn:
         latest_date = get_latest_date_from_zip_table(conn)
+    if not latest_date:
+        latest_date = datetime.datetime(year=2018, month=1, day=1)
     today = datetime.datetime.now()
+    if 'INIT_DB' in os.environ:
+        print('Running crawler in INIT_DB mode, downloading all data except for the last 14 days.')
+        today -= datetime.timedelta(days=14)
 
     from_datestr = latest_date.strftime('%Y%m%d')
     to_datestr = today.strftime('%Y%m%d')
 
-    res = subprocess.run([
-        'waybackpack',
-        '--list',
-        '--from-date', from_datestr,
-        '--to-date', to_datestr,
-        '--user-agent', 'waybackpack-1up+sdcounty@posteo.net',
-        COVID_BY_ZIP_URL
-    ], capture_output=True)
-    urls = res.stdout.decode().split()
-    urls = urls[:min([MAX_DOWNLOADS, len(urls)])]
-    print('Downloading {} files...'.format(len(urls)))
+    if not skip_download:
+        res = subprocess.run([
+            'waybackpack',
+            '--list',
+            '--from-date', from_datestr,
+            '--to-date', to_datestr,
+            '--uniques-only',
+            '--user-agent', 'waybackpack-1up+sdcounty@posteo.net',
+            COVID_BY_ZIP_URL
+        ], capture_output=True)
+        urls = res.stdout.decode().split()
+        if 'INIT_DB' not in os.environ:
+            urls = urls[:min([MAX_DOWNLOADS, len(urls)])]
+        print('Downloading {} files...'.format(len(urls)))
 
-    for url in urls:
-        try:
-            print('Downloading {}'.format(url))
-            response = requests_retry_session().get(url)
-        except Exception as x:
-            print('Download for {} failed:'.format(url), x.__class__.__name__)
-            continue
-        if response.status_code != 200:
-            print('Download for {} failed with status code {}.'.format(url, response.status_code))
-            continue
-        target_dir = data_dir.joinpath('./{}'.format(re.findall(r'/web/([0-9]+)/', url)[0]))
-        target_dir.mkdir()
-        with open(data_dir.joinpath('./{}/download.pdf'.format(re.findall(r'/web/([0-9]+)/', url)[0])), 'wb') as file:
-            file.write(response.content)
+        for url in urls:
+            try:
+                print('Downloading {}'.format(url))
+                response = requests_retry_session().get(url)
+            except Exception as x:
+                print('Download for {} failed:'.format(url), x.__class__.__name__)
+                continue
+            if response.status_code != 200:
+                print('Download for {} failed with status code {}.'.format(url, response.status_code))
+                continue
+            target_dir = data_dir.joinpath('./{}'.format(re.findall(r'/web/([0-9]+)/', url)[0]))
+            target_dir.mkdir()
+            with open(data_dir.joinpath('./{}/download.pdf'.format(re.findall(r'/web/([0-9]+)/', url)[0])), 'wb') as file:
+                file.write(response.content)
+    else:
+        print('Skipping download since files already exist in target folder.')
 
     with connection as conn:
-        for nx, timestamp_dir in enumerate(data_dir.glob('*')):
-            if nx == MAX_DOWNLOADS:
-                break
+        for timestamp_dir in sorted(data_dir.glob('*')):
             pdf_file = extract_pdf(timestamp_dir)
             header_data = PdfFileReader(open(pdf_file, 'rb')).getPage(0).extractText()
             data_raw = tabula.read_pdf(pdf_file, lattice=True, multiple_tables=False).to_csv()
@@ -100,8 +142,9 @@ if __name__ == '__main__':
                 start_row = ['Zip Code,Count' in x for x in data_raw.split('\n')].index(True)
             data = pd.read_csv(StringIO('\n'.join(data_raw.split('\n')[start_row:])))
             print(data.head(3))
-            if '1' in data.columns:
-                data = data.drop(columns=['1'])
+            for x in range(10):
+                if str(x) in data.columns:
+                    data = data.drop(columns=[str(x)])
             if 'Rate per 100,000*' in data.columns:
                 data = data.drop(columns=['Rate per 100,000*'])
 
@@ -123,6 +166,7 @@ if __name__ == '__main__':
                     col_data = col_data.iloc[:first_non_zip_nx, :]
                 except ValueError:
                     pass
+                col_data['cases'] = col_data['cases'].apply(lambda x: x.replace(',',''))
                 col_data = col_data.astype({'zip': int, 'cases': float})
                 col_data.insert(0, 'date', date)
                 return col_data
